requirements.txt
fastapi==0.110.0
uvicorn==0.17.0
python-multipart==0.0.5
boto3==1.21.0
python-dotenv==0.19.2
openai==1.14.2
pydantic==2.6.4
langchain==0.1.13
langchain-community==0.0.29